+++ cilk/fib_lib.h
@@ -2,7 +2,7 @@
 #define FIB_LIB_H
 #include <stddef.h>
 
-size_t fib_scope(int n);
-size_t fib_noscope(int n);
+size_t fib_scope(size_t n);
+size_t fib_noscope(size_t n);
 
 #endif // FIB_LIB_H
